early demo created
The model is very buggy, so the tracker isn't too great. Had to make the up angle 50, and down angle 140 instead of 30 and 160 just to give the model more leniency. Also only works with left curl. Future plans to deploy it with react, and make it so there are multiple options to test out.